fixed bug where AI could become the card czar
ManagePlayersActivity was sorting the player list, which was never
intended to be sorted since the card czar index remains the same and
would essentially make a new player the card czar.  To fix this,
ManagePlayersActivity works with a copy of the player array.

Added a menu option to perform a total game reset for debugging.